Optimised Registry::find_components by using std::views::filter and std::views::transform which returns an iterator instead of passing the vector.
Fixed security issue with test_map.cpp where multiplying two ints and them implicitly converting them to size_t could result in an arithmetic overflow.

The headless pyglet problem still exists however, but I'm not sure why.

+++ src/hades_extensions/tests/generation/test_map.cpp
@@ -228,7 +228,7 @@ TEST_F(MapFixture, TestMapCreateHallwaysNoConnections) {
 /// Test that running cellular automata on a grid with all empty tiles doesn't do anything.
 TEST_F(MapFixture, TestMapRunCellularAutomataAllEmpty) {
   run_cellular_automata(grid);
-  ASSERT_EQ(*grid.grid, std::vector(grid.width * grid.height, TileType::Empty));
+  ASSERT_EQ(*grid.grid, std::vector(static_cast<std::size_t>(grid.width * grid.height), TileType::Empty));
 }
 
 /// Test that running cellular automata on a grid with all floor tiles sets the edges to walls.
@@ -257,7 +257,7 @@ TEST_F(MapFixture, TestMapRunCellularAutomataAllWalls) {
     }
   }
   run_cellular_automata(grid);
-  ASSERT_EQ(*grid.grid, std::vector(grid.width * grid.height, TileType::Empty));
+  ASSERT_EQ(*grid.grid, std::vector(static_cast<std::size_t>(grid.width * grid.height), TileType::Empty));
 }
 
 /// Test that running cellular automata on a grid with mixed floor and wall tiles works correctly.
